src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["y2keylla <keylla.bispo@sou.inteli.edu.br>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.10.16"
jupyter = "^1.1.1"
pandas = "^2.2.3"
numpy = "^1.24.0"
boto3 = "^1.35.44"
pyarrow = "^16.0.0"
matplot = "^0.1.9"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
plotly = "^5.24.1"
python-dotenv = "^1.0.1"
uuid = "^1.30"
datetime = "^5.5"
skimpy = "^0.0.15"
openpyxl = "^3.1.5"
flask = "^3.0.3"
psycopg = "^3.2.3"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9"
flasgger = "^0.9.7.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
clickhouse-driver = "^0.2.9"
clickhouse-connect = "^0.8.6"
prefect = "^3.1.4"
streamlit = "^1.40.2"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-html = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
